+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Offline players now have their data cleared when joining the server after a reset (@leumasme)
 - More data is cleared on reset
+- Migrated commands to native brigadier for 
 
 ## [1.2.1] - 2024-07-22